fix test
* fix test

* fmt

* fix test

* fix test

* fix test

* fix unboud ong test

* fix eip155tx test

* fix deserialize bug

* update

* add api v2

* fix v2

* fmt code

* fmt code

+++ smartcontract/service/native/utils/unbind_ong_test.go
@@ -59,13 +59,13 @@ func TestCumulative(t *testing.T) {
 // test 1 balance will not get ONT_TOTAL_SUPPLY eventually
 func TestTotalONG(t *testing.T) {
 	total := CalcUnbindOng(constants.ONT_TOTAL_SUPPLY, 0, TIME_INTERVAL*18) + CalcGovernanceUnbindOng(0, TIME_INTERVAL*18)
-	assert.Equal(t, total, constants.ONG_TOTAL_SUPPLY)
+	assert.Equal(t, total, uint64(constants.ONG_TOTAL_SUPPLY))
 
 	total = CalcUnbindOng(constants.ONT_TOTAL_SUPPLY, 0, TIME_INTERVAL*108) + CalcGovernanceUnbindOng(0, TIME_INTERVAL*108)
-	assert.Equal(t, total, constants.ONG_TOTAL_SUPPLY)
+	assert.Equal(t, total, uint64(constants.ONG_TOTAL_SUPPLY))
 
 	total = CalcUnbindOng(constants.ONT_TOTAL_SUPPLY, 0, ^uint32(0)) + CalcGovernanceUnbindOng(0, ^uint32(0))
-	assert.Equal(t, total, constants.ONG_TOTAL_SUPPLY)
+	assert.Equal(t, total, uint64(constants.ONG_TOTAL_SUPPLY))
 }
 
 func TestCalcGovernanceUnbindOng(t *testing.T) {
